models/deepfake-img/requirements.txt
transformers>=4.49.0
torch
Pillow
gradio
